requirements.txt
django
djangorestframework
Pillow
psycopg2-binary
python-decouple
python-slugify
django-filter
stripe

openmeteo-requests
requests-cache retry-requests numpy pandas
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.14.0
idna==3.6
iniconfig==2.0.0
lxml==4.9.3
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.1
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.35.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
virtualenv==20.25.0
